+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/SolutionView.java
@@ -137,7 +137,7 @@ public interface SolutionView extends ProblemView {
         writeToFile(targetFolder.resolve("constraint.state.xml"), constraint().toDom());
         writeToFile(targetFolder.resolve("demands.fods"), demands().toFods());
         writeToFile(targetFolder.resolve("supplies.fods"), supplies().toFods());
-        writeToFile(targetFolder.resolve("history.fods"), history().toFods());
+        writeToFile(targetFolder.resolve("history.fods"), history().toAnalysisFods());
         writeToFile(targetFolder.resolve("constraint.natural-argumentation.txt")
                 , constraint()
                         .naturalArgumentation()